Add --tickets to munge more attachments that have encoded content
Previously attachments were pre-filtered by the search string at SQL
level, to limit the to-be-processed attachments to a reasonable small
set. The drawback is it bypassed attachments that have encodings like
base64, etc. even if the decoded content matches and should be replaced.
Because of this, we couldn't munge non-ascii strings as they are quite
probably encoded in the database.

To get around the pre-filter behavior and still keep to-be-processed
attachments being a small set, --tickets is added. Besides, in real
life, we believe it'll be the most common usage to munge attachments for
only a few tickets.